lab 2 done..

+++ tellez.aron/I2/largestdengthdecreasing.cpp
@@ -1,8 +1,8 @@
 #include "largestdengthdecreasing.h"
 size_t largest_dength_decreasing(const int *arr, const size_t size)
 {
-  int max = 0;
-  int len = 1;
+  size_t max = 0;
+  size_t len = 1;
   for (int i = 0; i < size - 1; i++)
   {
     if (arr[i + 1] < arr[i])
